Adjust AppAlert component and Login page layout
Previous Login page contained a background illustration of Admin-amico.svg, two columns and a form with an Alert component. The columns and form were placed side-by-side, with a max-width of 75% for the form. This code change replaces the previous layout with a single column, now containing the form, Alert component as a child of the form and the previous background illustration. The updated column occupies 8 out of 12 available column widths and there is an updated target for the Storyset link preview. Additionally, a left border was added to the Alert component and overflow was hidden for the container.